--- frontend/src/pages/oauth/code/index.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function index() {
-  return <div>ㅋㅋㅋ</div>;
-}
